[server] | feat(Map): 커스텀 맵 제작
Uploaded video info나 user session 정보를 서버에서 관리하기 위해 맵을 사용했습니다.
두 군데의 중복코드가 존재해 부모 Map을 만들어 자식에서 상속받아 쓰도록 했습니다
